+++ include/minml/utils/serialization.hpp
@@ -9,7 +9,6 @@
 #define MINML_UTILS_SERIALIZATION_HPP_
 
 #include <cstring>
-#include <stdexcept>
 #include <string>
 #include <vector>
 
@@ -18,11 +17,11 @@ class SerializedObject
 	private:
 		std::vector<uint8_t> m_data;
 	public:
-		explicit SerializedObject(size_t size = 0);
+		explicit SerializedObject(int64_t size = 0);
 		SerializedObject(const std::string &path);
 
-		size_t size() const noexcept;
-		size_t capacity() const noexcept;
+		int64_t size() const noexcept;
+		int64_t capacity() const noexcept;
 		void clear() noexcept;
 
 		void saveToFile(const std::string &path) const;
@@ -31,16 +30,16 @@ class SerializedObject
 		const uint8_t* data() const noexcept;
 		uint8_t* data() noexcept;
 
-		void save(const void *src, size_t size_in_bytes);
-		void load(void *dst, size_t offset, size_t size_in_bytes) const;
-		
+		void save(const void *src, int64_t size_in_bytes);
+		void load(void *dst, int64_t offset, int64_t size_in_bytes) const;
+
 		template<typename T>
 		void save(const T &value)
 		{
 			this->save(&value, sizeof(T));
 		}
 		template<typename T>
-		T load(size_t offset) const
+		T load(int64_t offset) const
 		{
 			T result;
 			this->load(&result, offset, sizeof(T));
@@ -48,7 +47,7 @@ class SerializedObject
 		}
 
 		//common to all classes
-		size_t getMemory() const noexcept;
+		int64_t getMemory() const noexcept;
 };
 
 #endif /* MINML_UTILS_SERIALIZATION_HPP_ */
